% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mm_aic}
\alias{mm_aic}
\title{Compute AIC for fitted model}
\usage{
mm_aic(fit)
}
\arguments{
\item{fit}{An mm_fit object}
}
\value{
Numeric AIC value
}
\description{
Compute AIC for fitted model
}
\examples{
\dontrun{
fit <- mm_gmm_fit(iris[,1:4], k = 3)
mm_aic(fit)
}
}
