% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mm_bic}
\alias{mm_bic}
\title{Compute BIC for fitted model}
\usage{
mm_bic(fit)
}
\arguments{
\item{fit}{An mm_fit object}
}
\value{
Numeric BIC value
}
\description{
Compute BIC for fitted model
}
\examples{
\dontrun{
fit <- mm_gmm_fit(iris[,1:4], k = 3)
mm_bic(fit)
}
}
