% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{mm_conda_run}
\alias{mm_conda_run}
\title{Run a conda command (stops on error, prints output)}
\usage{
mm_conda_run(args)
}
\description{
Run a conda command (stops on error, prints output)
}
\keyword{internal}
