% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{mm_ensure_miniconda}
\alias{mm_ensure_miniconda}
\title{Ensure Miniconda/Miniforge is installed and discoverable}
\usage{
mm_ensure_miniconda()
}
\description{
Ensure Miniconda/Miniforge is installed and discoverable
}
\keyword{internal}
