% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mm_likelihood}
\alias{mm_likelihood}
\title{Compute log-likelihood for fitted model}
\usage{
mm_likelihood(fit)
}
\arguments{
\item{fit}{An mm_fit object}
}
\value{
Numeric log-likelihood value
}
\description{
Compute log-likelihood for fitted model
}
\examples{
\dontrun{
fit <- mm_gmm_fit(iris[,1:4], k = 3)
mm_likelihood(fit)
}
}
