% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{mm_miniconda_installed}
\alias{mm_miniconda_installed}
\title{Is Miniconda/Miniforge installed (reticulate-managed)?}
\usage{
mm_miniconda_installed()
}
\description{
Is Miniconda/Miniforge installed (reticulate-managed)?
}
\keyword{internal}
