% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mm_new_fit}
\alias{mm_new_fit}
\title{Create a new mm_fit object}
\usage{
mm_new_fit(
  py_model,
  params_store,
  final_params,
  model_name,
  k,
  call,
  n,
  d,
  optimizer,
  converged = NULL
)
}
\description{
Create a new mm_fit object
}
\keyword{internal}
