% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mm_params}
\alias{mm_params}
\title{Get parameter values from fit object}
\usage{
mm_params(fit, convert = TRUE)
}
\arguments{
\item{fit}{An mm_fit object}

\item{convert}{Logical, whether to convert Python objects to R (default TRUE)}
}
\value{
Parameter object (structure depends on model family)
}
\description{
Extracts the parameter values from a fitted mixture model.
}
\examples{
\dontrun{
fit <- mm_gmm_fit(iris[,1:4], k = 3)
params <- mm_params(fit)
}
}
