% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{mm_pip_install}
\alias{mm_pip_install}
\title{Run pip install in the currently active Python}
\usage{
mm_pip_install(packages, extra_args = character())
}
\arguments{
\item{packages}{character vector of requirement strings}

\item{extra_args}{character vector of extra pip args (e.g., "--no-deps")}
}
\description{
Run pip install in the currently active Python
}
\keyword{internal}
