% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{mm_py_initialized}
\alias{mm_py_initialized}
\title{Is reticulate Python already initialized in this R session?}
\usage{
mm_py_initialized()
}
\description{
Is reticulate Python already initialized in this R session?
}
\keyword{internal}
