% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{mm_python_available}
\alias{mm_python_available}
\title{Check if Python module is available}
\usage{
mm_python_available()
}
\value{
Logical indicating if either mixture_models or Mixture_Models is importable
}
\description{
Check if Python module is available
}
