% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{mm_require_python}
\alias{mm_require_python}
\title{Provision Python environment with required packages}
\usage{
mm_require_python(force = FALSE)
}
\arguments{
\item{force}{Logical, force re-provisioning}
}
\description{
Provision Python environment with required packages
}
\keyword{internal}
