% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{mm_setup_conda}
\alias{mm_setup_conda}
\title{Provision a conda env (Python 3.10 + NumPy 1.23.5) and install Mixture-Models}
\usage{
mm_setup_conda(force = FALSE)
}
\arguments{
\item{force}{Logical, remove and recreate the conda env}
}
\description{
Provision a conda env (Python 3.10 + NumPy 1.23.5) and install Mixture-Models
}
\keyword{internal}
