% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{mm_use_env}
\alias{mm_use_env}
\title{Use configured Python environment (if already available)}
\usage{
mm_use_env(required = FALSE)
}
\arguments{
\item{required}{Logical, whether to require Python availability}
}
\description{
Use configured Python environment (if already available)
}
\keyword{internal}
