% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.mm_fit}
\alias{print.mm_fit}
\title{Print method for mm_fit objects}
\usage{
\method{print}{mm_fit}(x, ...)
}
\arguments{
\item{x}{An mm_fit object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for mm_fit objects
}
