% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{mm_require_python}
\alias{mm_require_python}
\title{Ensure Python is configured for this session}
\usage{
mm_require_python(force = FALSE)
}
\arguments{
\item{force}{Logical, force re-provisioning if using conda env}
}
\description{
Ensure Python is configured for this session
}
\keyword{internal}
